refactor and add FileSystemStorage
